--- src/components/LandingComponents/Footer/Footer.module.css
@@ -1,79 +1,0 @@
-.footer {
-    color: #B2BBCC;
-    font-family: "Poppins";
-}
-
-.footerBoxContainer {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 2rem;
-}
-
-.footerBox {
-    flex: 1 1 25rem;
-}
-
-.footerBox h3 {
-    color: #ffffff;
-    font-weight: 600;
-    font-size: 1.5rem;
-    line-height: 2rem;
-    padding: 1rem 0;
-    font-family: "Poppins" !important;
-}
-
-.footerBanner {
-    background: linear-gradient(91.94deg, #B17E32 -8.77%, #F3E681 58.63%, #CF9F41 107.45%);
-    height: 0.25rem;
-    width: 1.875rem;
-    border-radius: 0.125rem;
-    margin-top: 0ch !important;
-    padding-top: 0ch !important;
-    margin-bottom: 1rem;
-}
-
-.footerBox p {
-    font-size: 1.2rem;
-}
-
-.footerBox a {
-    font-family: "Poppins" !important;
-    font-weight: 500;
-    font-size: 1.2rem;
-    line-height: 2rem;
-    display: block;
-    color: #B2BBCC;
-    padding: 0.5rem 0;
-    text-decoration: none;
-}
-
-.footerBox a:hover {
-    color: #fff;
-}
-
-.footerBox img {
-    margin-top: 1rem;
-}
-
-.footerMedia {
-    display: flex;
-    gap: 0.625rem;
-    font-size: 1.5rem;
-}
-
-.footerCredit {
-    text-align: center;
-    margin-top: 1.5rem;
-    padding-top: 2.5rem;
-    font-size: 1rem;
-    font-family: "Poppins";
-    font-weight: 400;
-    line-height: 1.125rem;
-    color: #8C9AB4;
-    border-top: 0.1rem solid rgba(255, 255, 255, 0.1);
-    display: flex;
-    gap: 1rem;
-    justify-content: space-between;
-    align-items: center;
-    flex-wrap: wrap;
-}
