+++ src/landing/header/Header.module.css
@@ -1,0 +1,208 @@
+.header {
+  height: 12vh;
+  position: sticky;
+  top: 0;
+  left: 0;
+  right: 0;
+  z-index: 1000;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background-color: #252524;
+  padding: 0rem 9%;
+}
+
+#menuBtn {
+  cursor: pointer;
+  font-size: 3.5rem;
+  display: none !important;
+  margin-right: 1rem;
+  color: #fff;
+}
+
+.headerLogo {
+  margin-left: 1rem;
+  font-family: "Popins", "Nunito Sans", sans-serif;
+  font-size: 2.5rem;
+  font-weight: 700;
+  line-height: 3.25rem;
+  color: white;
+  text-decoration: none;
+  text-wrap: nowrap;
+}
+
+.navbar {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  align-items: center;
+  gap: 1rem;
+}
+
+.navbar a {
+  position: relative;
+  margin: 0 1rem;
+  font-family: "Poppins", "Nunito Sans", sans-serif;
+  font-size: 1.75rem;
+  font-weight: 500;
+  line-height: 1.1875rem;
+  text-decoration: none;
+  color: #fff;
+  text-wrap: nowrap;
+  transition: all 0.3s ease;
+  background: none;
+}
+
+.navbar a::after {
+  content: "";
+  position: absolute;
+  left: 0;
+  bottom: -0.2rem;
+  width: 0%;
+  height: 2px;
+  background: linear-gradient(90deg, #cf9f41 0%, #f3e681 50%, #b17e32 100%);
+  transition: width 0.4s ease;
+}
+
+.navbar a:hover {
+  background: linear-gradient(90deg, #ffd700, #ffea82);
+  background-clip: text;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  transform: translateY(-5px);
+}
+
+.navbar a:hover::after {
+  width: 100%;
+}
+
+.navbar .activeLink {
+  color: #ffd700;
+}
+
+.headerBtns {
+  margin: 0.625rem;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.625rem;
+  align-items: center;
+}
+
+.btnLogin {
+  cursor: pointer;
+  background-color: #252524;
+  width: 10.4375rem;
+  height: 3rem;
+  border: 0;
+  border-radius: 6.25rem;
+  font-family: "Poppins", "Nunito Sans";
+  font-weight: 400;
+  font-size: 1.3rem;
+  line-height: 1.5rem;
+  text-align: center;
+  color: white;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  box-sizing: border-box;
+}
+
+.btnSignup {
+  cursor: pointer;
+  background: linear-gradient(90deg, #b17e32 0%, #f3e681 50%, #cf9f41 100%);
+  width: auto;
+  height: 3.5rem;
+  border-radius: 50px;
+  padding: 0.75rem 1.5rem;
+  font-family: "Poppins", "Nunito Sans", sans-serif;
+  font-weight: 500;
+  font-size: 1.3rem;
+  line-height: 1.5rem;
+  text-align: center;
+  color: #242424;
+  display: inline-flex;
+  justify-content: center;
+  align-items: center;
+  box-sizing: border-box;
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.15);
+  transition: all 0.9s ease;
+}
+
+.btnSignup:hover {
+  background: linear-gradient(90deg, #cf9f41 0%, #f3e681 50%, #b17e32 100%);
+  box-shadow: 0 12px 24px rgba(0, 0, 0, 0.25);
+  transform: translateY(-3px);
+  color: #ffffff;
+}
+
+.btnSignup:active {
+  transform: translateY(0);
+  box-shadow: 0 6px 12px rgba(0, 0, 0, 0.15);
+}
+
+/* 1200, max width  */
+@media (max-width: 1200px) {
+  .header {
+    padding: 0rem 2rem;
+    /* 0px 32px */
+  }
+
+  .header .headerBtns {
+    display: none !important;
+  }
+}
+
+/* 768 max width  */
+@media (max-width: 768px) {
+  .header {
+    height: auto;
+    min-height: 12vh;
+    justify-content: flex-start;
+    align-items: center;
+    padding: 0rem 2rem;
+  }
+
+  #menuBtn {
+    display: inline-block;
+    -webkit-transition: 0.2s linear;
+    transition: 0.2s linear;
+  }
+
+  #menuBtn.fa-times {
+    -webkit-transform: rotate(180deg);
+    transform: rotate(180deg);
+  }
+
+  .header .btn {
+    display: none;
+  }
+
+  .header .navbar {
+    position: absolute;
+    top: 99%;
+    left: 0;
+    right: 0;
+    border-top: 0.2rem solid #000;
+    /* 3.2px */
+    background: #000;
+    -webkit-clip-path: polygon(0 0, 100% 0, 100% 0, 0 0);
+    clip-path: polygon(0 0, 100% 0, 100% 0, 0 0);
+    -webkit-transition: 0.2s linear;
+    transition: 0.2s linear;
+  }
+
+  .header .navbar .active {
+    -webkit-clip-path: polygon(0 0, 100% 0, 100% 100%, 0 100%);
+    clip-path: polygon(0 0, 100% 0, 100% 100%, 0 100%);
+    flex-direction: column;
+    align-items: flex-start;
+    align-content: flex-start;
+  }
+
+  .header .navbar a {
+    display: block;
+    margin: 0;
+    padding: 1.5rem 2rem;
+    /* 24px 32px */
+  }
+}
